Removed unnecessary offset time calculations
1. Fixed logic
2. Fixed tests

The `offsetTime` should not affect the request to the calendar API.
`offsetTime` should include a time offset from the beginning of the period.

The full implementation of `offsetTime` will be taken into account by https://issues.folio.org/browse/CIRC-160
In the current implementation there is a small possibility of time offset.