Instead of returning null, return an HttpResponse
It was rightfully pointed out to me that my hasty fix lacked quite a bit. I was not too thrilled about returning `null` to indicate a specific failure, so I agreed. The better approach, thanks to Marc, is to return an HttpResult. Now we do that. Now the code flows much better. The error message could be better. :)