CIRC-440: prevent multiple recalls from changing the loan due date
The loan due date should only be changed once due to a recall. Subsequent recalls should not modify the due date. We will now set and respect the `dueDateChangedByRecall` flag. It will be set on a recall, if it has not already been set, and when it comes time to apply recall logic to update the loan, if this flag is set, we no longer perform the recall logic.

This fix also has changed an expectation of CIRC-289 (moving recall requests) and the unit test was updated to reflect the new expected outcome.