Fix for CIRC-463 - request position updates in the queue
In order to ensure that we comply with DB constraints, we need to ensure
we process the request updates in an order fashion, one after the other.
The problem here is that we execute the updates asynchronously, without
waiting for the result of the previous update. In this case, we can have
requests being updated to positions that have not been vacated.

To address this, we now wait to issue the next update until after the
previous update has returned. This will prevent, or, at least, severely
reduce the chance of a request being moved to an already taken position.

Also added a unit test that fails with the existing code and passes with
this fix. In order to mock this, added a test dependency to Mockito and
made the `RequestQueueRepository` constructor package-private. Mocking
static methods is more challenging and typically requires more
heavy-weight mocking frameworks. Typically, package-private constructors
are acceptable.

It is not clear that this unit test is really required. I was just using
TDD here to prove what I changed fixed the problem and I thought it
might be useful to prevent regressions. We can drop the test and changes
associated with it if it is too problematic.