CIRC-367: Anonymize closed loans setting of "X interval after loan closes"
The loan anonymization process is scheduled to run in short intervals; every minute (set in the module-descriptor) okapi makes a call to /circulation/scheduled-anonymize-processing, which kicks off the loan anonymization in the current tenant. It finds all closed loans and anonymizes them based on the loan history settings. Fees and fines settings are to be followed regardless of whether it differs from the general setting for anonymization.
The closing date of a fee/fine is not saved as a field, so is derived from the feefineactions of that fee/fine.