Verbose explanation of criterium syntax, complete example, heading for "!" and "all"
The current documentation about the criterium syntax is illegible, it hides the syntax,
and we cannot expect people to read the document but we must assume that they only
skim through it.

Therefore the explanation needs to be verbose by using complete examples that stands
out instead of the in-line examples.

A heading needs to be added for the "!" and "all" syntax to catch the eye.