CIRC-637: replace thenCombine() logic with thenApply()
While `thenCombine()` is preferable, the 404 from non-existent
loans will bubble down the chain. It is prefered that every error,
except for 404 on existing loans, should bubble down.

I do not (yet) know of any way to selectively bubble down specific
failures, so redesign this to use the thenApply().

Exceptions on loading any existing Loan will now be silently ignored
and the is claimed returned due date test will not be performed.

In addition, do not perform the is claimed returned due date test
when either upcoming or previous are NULL instead of just when
upcoming is NULL.