CIRC-647: Send notice on manual due date change
As noted in CIRC-439, a manual change to the due date should send a notice
when the notice emails for this kind of change are enabled.

This makes some changes to use Loan instead of LoanAndRelatedRecords.
That set of changes should be considered and either kept or these changes
should be refactored to utilize LoanAndRelatedRecords.

Warning: tests are currently failing in as far as I can tell a nonsensical way.
The tests are saying that two strings are not matching, provides those two
supposedly different strings, and it turns out those two strings are actually
identical when visually comparing them.

Additional Changes will follow this to address the test failures.

see: https://issues.folio.org/browse/CIRC-439