Refactored Resources to consolidate duplicate code
Moved Note creation code to an abstract resource class.

The DeclareLostResource needs to create notes only when the item
was already claimed returned and was doing this before the item was
actually declared lost. I was able to set aside if a note should be
created and now the note is only created if it was claimed returned
and the item was succsessfully declared lost.