CIRC-940: change approach for fetching, add missing functionality.
Get Loans and then Items instead of Items and then Loans.
Doing this results in a substantial amount of changes.

There were several things missed, such as the permissions.

Better prove a "skeleton" on the return instead of the entire Loan and Item.

More CQL functions are necessary to wrap conditions in parenthesis.

There are some FIXME: in this commit.
These cases are to be fixed once the situation with "Aged to Lost" with or without "Set Cost" is addressed.