CIRC-995: Improve tests.
The tests that I wrote passed, but I completely forgot to run all the tests.
This revealed that I shouldn't be using join() the way that I was using it.
Redesign this to avoid directly calling join().

After running additional all the tests, I can see that there is more work to do.
This only resolves the join() problem with the current tests.