CIRC-1000: Do not charge overdue fines if lost item returned/renewed after date lost item fees may be refunded
The fix for this was relatively simple:  I removed the code that checks
to see if a fee or fine has already been applied when deciding wether
overdues should be charged for lost items.  that section of code already
checks to see if the item has a policy that charges overdues attached,
and wether the refund date has passed.  Without the check for existing
charges, the remaining code produces the specified behavior without
further alteration.

That was the easy part.

Testing for this behavior was...complicated.  It requires simulating a loan
with VERY specific policy settings, and the loans have to be artificially
aged to very specific points in time, then a checkin or renewal event had
to be applied to the loan at a specific point in time as well.  Many existing
tests test for conditions somewhat like this, but none checked for the very
specific combination of conditions needed.  I had to make some alterations to
several fixture files, most notably AgeToLostFixture, so that I could create
loans with the correct policy settings.  I wrote four tests: two check the
behavior for items checked in after the refund date has passed (one for aged
to lost and one for declared lost), and two that check for behavior when a
renewal override is performed (again, one for aged-to-lost, and one for
declared lost).  I tried to place tests in files where I had access to
classes and context that I could use to construct the tests, but also where
they seemed to make logistical sense.  Tests using renewal override I put in
the file with other renewal override tests, for example.  As I was writing
them I often had the code dump the entire loan object produced to screen so
that I could make sure the policies were correctly attached, the dates were
correct, etc.

refs: https://issues.folio.org/browse/CIRC-1000