CIRC-1086: XStream 1.4.15 Arbitrary File Deletion (CVE-2020-26259)
XStream is vulnerable to an Arbitrary File Deletion on the local host
when unmarshalling as long as the executing process has sufficient rights.

All versions until and including version 1.4.14 are affected running in a
Java environment containing the JAX-WS runtime, if using the version out of
the box.

Details: https://x-stream.github.io/CVE-2020-26259.html

Fix:

Update the XStream version from 1.4.14 to 1.4.15.