CIRC-966: Stage 2, Implement basic ClockManager and DateTimeUtil.
There already is a ClockManager and a DateTimeUtil.
Update these existing classes as a starting point for the planned changes necessary to migrate from JodaTime to JavaTime.

Of particular note, the existing behavior instantiates a static class of itself and then passes itself along via a static method.
I do not understand the reasoning behind this.
I opted to just have the `Clock` statically defined and this `Clock` will be passed to JavaTime.
This `Clock` is also used to generate the times used by the JodaTime implementation.
The `ClockManager` is mostly being used by the tests.

There are many methods implemented in these in a fassion that allows for a staged approach.
Many of these, now documentated, methods will be removed, changed, or replaced by the time JodaTime is fully replaced with JavaTime.

The ultimate goal is to have most, if not all, of the date and time functionality be accessed through either the `ClockManager` or the `DateTimeUtil`.

Another important note is that the previous code is already intermixing JodaTime with JavaTime.
As a result of the precarious situation, the `ClockManager` and `DateTimeUtil` currently has a mixture of JavaTime and JodaTime.

There is also the existing `ClockManager.getDateTime()` method.
The previous behavior is o get the time from `ZonedDateTime.now()` and then cast to a `DateTime` with the timezone set to UTC.
This behavior is maintained for compatibility but it seems like setting the time zone from the `Clock` is more appropriate.