CIRC-966: Stage 3, For CU & DTU, implement and refactor now() and similar.
Refactor `now()` and similar calls to use the appropriate `ClockUtil` or `DateTimeUtil` methods.

Fix several tests where the clock is not being correctly adjusted.

Fix one test where test is incorrect (originally fixed here: 69ad724614b84a8d9bdeb5ff72727047fc47f1a2).

Always reset the Clock at the end of changing it (such as with `@AfterEach`).
This solves a problem where the changing statically defined Clock in one test could result in the altered Clock being used in a completely different test.

The `runWithFrozenTime()` method is redundant and now wrong.
It is now wrong because it sets the JodaTime internal clock, which is no longer being used as of this commit and going forward.
Remove runWithFrozenTime()` entirely.
Remove `executeWithFixedDateTime()` entirely, for the same reasons.

The matcher should truncate both dates to milliseconds rather than just one of the dates.