CIRC-966: Stage 4, For CU & DTU, start of day and end of day methods.
The start and end of day utility methods are implemented along with their respective tests and are in use.
I discovered that there are also start/end of year and start/end of month.
These utility methods have also been implemented.

The start/end year/month/day methods are implemented to accept a zone (or default to a zone on an Object if available) in order to ensure that the order in which the zone is applied is consistent and correct.
If the time zone is applied at the wrong step, then the start and end times will be incorrect.

The `Request.truncateRequestExpirationDateToTheEndOfTheDay()` method is calling `.withZoneRetainFields()` when it should instead be calling `.withZone()`.
The related test also seems wrong.
Both of these have been corrected.

There is a problem with `KeepCurrentDateStrategy.calculateDueDate()` where the requested date is probably incorrectly altered.
Based on UICIRC-565 and PR 759, this was done to solve a problem with incorrect date being displayed.
I believe that the problem is instead related to how the UI passes down a date with time in the tenants timezone for some date only value.
That date only value (date without time) is represented in the tenants timezone and is not directly related to the datetime representation.
This can be a bit confusing, so a test `shouldNeverChangeLocalDayWhenConvertingToTimeZone()` has been added to show this and the other test `shouldAlwaysKeepCurrentDateWhenConvertingToTimeZone()` has been disabled.
The fix in CIRC-565 is effectively reverted and that problem may re-appear with this commit.

see: https://issues.folio.org/browse/UICIRC-565
see: https://github.com/folio-org/mod-circulation/pull/759