CIRC-966: Stage 4, ensure date in UTC timezone is returned.
After some extensive discussion I was asked to confirm the following is the flow for `KeepCurrentDateStrategy.calculateDueDate()`:
1) Convert due date to tenant local time zone (without changing the actual point in time).
2) Move it to the end of the day.
3) Convert it back to UTC.

The third step "Convert it back to UTC", I was thinking that the date is relative to UTC.
It later occured to me that maybe it is actually that it just be returned in the UTC timezone.

To properly achieve this, one more step is needed.
Converting the timezone back to UTC.

The logic is now:
To calculate the end of the day, the calculation must be done in the tenants time zone.
Once the date is converted to the appropriate time zone and the end of the day is calculated, convert that result back to the UTC time zone.

This commit performs the last step by converting the date time into the UTC time zone.
The test is updated accordingly to expect a result with the time zone in UTC.