CIRC-966: Stage 7, JodaTime to JavaTime conversion.
Convert JodaTime to JavaTime, completely removing JodaTime from the dependency hierarchy.

There are multiple "Period" classes, one being from JodaTime.
The JavaTime change or removed the Period class, forcing significant changes in this regard.
There are quick and simple functions added, `plusDate()` and `minusDate()` as a patch to get this working for now.
There are likely much better ways to implement this using the Chrono* classes from JavaTime but the changeset would be larger than desired for this set of changes.
Some of the JodaTime Period functionality is replaced with the JavaTime Duration.

There are numerous inconsistencies in how truncation occurs on dates and times.
This attempts to preserve the original behavior as closely as possible.
For example, some of the date comparison functions are truncating to minutes and ignoring the seconds.

There no longer is an Interval class in JavaTime.
A custom class is implemented.
This class is implemented in a simple way and likely has room improvement.
Again, the goal is to limit the size of the changes and to keep any potential breaking changes as close to 0 as possible.

Despite the planning, there are still additional changes need in the utilities.

There are many cases where `int` is being used for date and time related calculations.
JavaTime uses `long` in many places.
To prevent problems, many of these `int` where changed into `long`.

There are many places in the code where MIDNIGHT is being used to represent to start of the day or is being used to derive the end of the day.
This is all well and good except there are odd cases, often being related to daylight savings time, that result in days not always starting at midnight.
This does not go as far as to correct all of these cases.
Anywhere that the values needed changes were changed to use the `atStartOfDay` or `atEndofDay` utility methods.

There are problems with `CheckOutCalculateDueDateShortTermTests.testMoveToTheEndOfCurrentServicePointHoursRolloverScenario` and `EndOfCurrentHoursStrategy.isDateEqualToBoundaryValueOfDay`.

The use of `OffsetDateTime` is removed, being replaced with `ZonedDateTime`.
Staying with `ZonedDateTime` is safer and still handles offsets.
Related utility methods are removed.

Behavioral changes are introduced in src/main/java/org/folio/circulation/domain/Request.java for truncateRequestExpirationDateToTheEndOfTheDay.
and als in src/main/java/org/folio/circulation/domain/policy/library/KeepCurrentDateStrategy.java for calculateDueDate().

All temporary JodaTime to JavaTime helper utility methods are now removed.