CIRC-966: Stage 7, remove dead code, improve test coverage.
The `defaultToClockZone()` call in `DateFormatUtil` does nothing given the wrapping null check.

After reviewing the coverage, handle some corner cases that were not handled.

There is still room for improvement in functions like `parseTime()` where the final return is not tested.
After looking at the code, I am not sure the best way to test it.
I might even be better to redesign the code not to return in this way because I suspect that invalid strings that fail the formatters will always throw an exception.
And the list the for loop is looping over is statically defined and should always return a size greater than zero.