CIRC-2210: Remove useless excessive INFO logging of rules in ExecutableRules
https://folio-org.atlassian.net/browse/CIRC-2210

ExecutableRules.determinePolicy runs

```
      log.info("Executing circulation rules: `{}` with parameters: `{}` to determine {}",
        text, parameters, policyType);
```

which prints the text variable that contains the circulation rules file that may contain hundres of lines.

This is printed whenever circulation rules are applied, this is very often.

This bloats the log.

This is not needed at INFO level.

And the log.info it is not needed at DEBUG level because the calling methods in ExecutableRules already log parameters and policyType, and CirculationRulesCache already logs the the rules when needed:

```
      log.info ("getRulesFromCache:: cache hit for tenant {}", tenantId);
      log.debug("getRulesFromCache:: cached rules: {}", cachedRules::getRulesAsText);
```

Task: Delete log.info in ExecutableRules.determinePolicy.